Get series values

The function get_series() is the main function of this package. Its goal is to facilitate direct access to the data published in the Atlas do Estado Brasileiro so that the user can work with them as they wish.

Usage

get_series(series_id = NULL, series_title = NULL)

Arguments

series_id, series_title

The series ids or titles to download

Value

A data.frame or a list containing the data from the series

Details

If the parameter used is for just one series, the result will be a data.frame containing the requested information. Now, if the parameter refers to more than one series, the return will be a list of data.frames, with each data.frame corresponding to a series.

Examples

get_series(series_id = 230)
#>      ano nivel_federativo cor rem_media nome_cor nome_nivel_federativo
#> 1   2004                E   1  4519.025 Indígena              Estadual
#> 2   2004                E   2  4031.816   Branca              Estadual
#> 3   2004                E   4  3691.235    Preta              Estadual
#> 4   2004                E   6  4256.041  Amarela              Estadual
#> 5   2004                E   8  3642.660    Parda              Estadual
#> 6   2004                F   1  9856.214 Indígena               Federal
#> 7   2004                F   2  9200.807   Branca               Federal
#> 8   2004                F   4  6265.668    Preta               Federal
#> 9   2004                F   6  9649.170  Amarela               Federal
#> 10  2004                F   8  7216.168    Parda               Federal
#> 11  2004                M   1  1790.644 Indígena             Municipal
#> 12  2004                M   2  2491.084   Branca             Municipal
#> 13  2004                M   4  1846.315    Preta             Municipal
#> 14  2004                M   6  2120.959  Amarela             Municipal
#> 15  2004                M   8  1773.207    Parda             Municipal
#> 16  2005                E   1  4937.175 Indígena              Estadual
#> 17  2005                E   2  4203.744   Branca              Estadual
#> 18  2005                E   4  3756.337    Preta              Estadual
#> 19  2005                E   6  4220.798  Amarela              Estadual
#> 20  2005                E   8  3783.586    Parda              Estadual
#> 21  2005                F   1 10158.056 Indígena               Federal
#> 22  2005                F   2  9670.056   Branca               Federal
#> 23  2005                F   4  6557.583    Preta               Federal
#> 24  2005                F   6 10039.441  Amarela               Federal
#> 25  2005                F   8  7506.265    Parda               Federal
#> 26  2005                M   1  1887.607 Indígena             Municipal
#> 27  2005                M   2  2573.102   Branca             Municipal
#> 28  2005                M   4  1874.541    Preta             Municipal
#> 29  2005                M   6  2157.068  Amarela             Municipal
#> 30  2005                M   8  1838.712    Parda             Municipal
#> 31  2006                E   1  5726.080 Indígena              Estadual
#> 32  2006                E   2  4575.902   Branca              Estadual
#> 33  2006                E   4  4087.969    Preta              Estadual
#> 34  2006                E   6  4460.946  Amarela              Estadual
#> 35  2006                E   8  4188.740    Parda              Estadual
#> 36  2006                F   1 11516.015 Indígena               Federal
#> 37  2006                F   2 10934.073   Branca               Federal
#> 38  2006                F   4  7403.890    Preta               Federal
#> 39  2006                F   6 11235.964  Amarela               Federal
#> 40  2006                F   8  8459.459    Parda               Federal
#> 41  2006                M   1  2070.473 Indígena             Municipal
#> 42  2006                M   2  2741.615   Branca             Municipal
#> 43  2006                M   4  2027.841    Preta             Municipal
#> 44  2006                M   6  2325.640  Amarela             Municipal
#> 45  2006                M   8  2015.738    Parda             Municipal
#> 46  2007                E   1  6292.626 Indígena              Estadual
#> 47  2007                E   2  4814.779   Branca              Estadual
#> 48  2007                E   4  4247.193    Preta              Estadual
#> 49  2007                E   6  4694.247  Amarela              Estadual
#> 50  2007                E   8  4358.516    Parda              Estadual
#> 51  2007                F   1 10982.924 Indígena               Federal
#> 52  2007                F   2 10549.032   Branca               Federal
#> 53  2007                F   4  7316.424    Preta               Federal
#> 54  2007                F   6 11457.370  Amarela               Federal
#> 55  2007                F   8  8102.741    Parda               Federal
#> 56  2007                M   1  2197.628 Indígena             Municipal
#> 57  2007                M   2  2841.842   Branca             Municipal
#> 58  2007                M   4  2135.360    Preta             Municipal
#> 59  2007                M   6  2444.818  Amarela             Municipal
#> 60  2007                M   8  2124.567    Parda             Municipal
#> 61  2008                E   1  6546.984 Indígena              Estadual
#> 62  2008                E   2  4900.346   Branca              Estadual
#> 63  2008                E   4  4293.029    Preta              Estadual
#> 64  2008                E   6  4684.763  Amarela              Estadual
#> 65  2008                E   8  4449.904    Parda              Estadual
#> 66  2008                F   1 12488.181 Indígena               Federal
#> 67  2008                F   2 11884.186   Branca               Federal
#> 68  2008                F   4  8134.122    Preta               Federal
#> 69  2008                F   6 12684.177  Amarela               Federal
#> 70  2008                F   8  9163.278    Parda               Federal
#> 71  2008                M   1  2324.924 Indígena             Municipal
#> 72  2008                M   2  2960.943   Branca             Municipal
#> 73  2008                M   4  2236.604    Preta             Municipal
#> 74  2008                M   6  2560.729  Amarela             Municipal
#> 75  2008                M   8  2240.009    Parda             Municipal
#> 76  2009                E   1  6688.836 Indígena              Estadual
#> 77  2009                E   2  5056.927   Branca              Estadual
#> 78  2009                E   4  4359.774    Preta              Estadual
#> 79  2009                E   6  4844.675  Amarela              Estadual
#> 80  2009                E   8  4649.487    Parda              Estadual
#> 81  2009                F   1 13454.535 Indígena               Federal
#> 82  2009                F   2 12867.808   Branca               Federal
#> 83  2009                F   4  8862.802    Preta               Federal
#> 84  2009                F   6 14146.405  Amarela               Federal
#> 85  2009                F   8  9926.399    Parda               Federal
#> 86  2009                M   1  2409.491 Indígena             Municipal
#> 87  2009                M   2  3069.064   Branca             Municipal
#> 88  2009                M   4  2326.254    Preta             Municipal
#> 89  2009                M   6  2655.172  Amarela             Municipal
#> 90  2009                M   8  2326.014    Parda             Municipal
#> 91  2010                E   1  7168.631 Indígena              Estadual
#> 92  2010                E   2  5324.013   Branca              Estadual
#> 93  2010                E   4  4562.799    Preta              Estadual
#> 94  2010                E   6  5217.690  Amarela              Estadual
#> 95  2010                E   8  4939.888    Parda              Estadual
#> 96  2010                F   1 13001.021 Indígena               Federal
#> 97  2010                F   2 12014.192   Branca               Federal
#> 98  2010                F   4  8482.114    Preta               Federal
#> 99  2010                F   6 13947.566  Amarela               Federal
#> 100 2010                F   8  9209.543    Parda               Federal
#> 101 2010                M   1  2469.880 Indígena             Municipal
#> 102 2010                M   2  3180.636   Branca             Municipal
#> 103 2010                M   4  2439.870    Preta             Municipal
#> 104 2010                M   6  2773.909  Amarela             Municipal
#> 105 2010                M   8  2445.325    Parda             Municipal
#> 106 2011                E   1  7213.093 Indígena              Estadual
#> 107 2011                E   2  5489.937   Branca              Estadual
#> 108 2011                E   4  4664.198    Preta              Estadual
#> 109 2011                E   6  5396.970  Amarela              Estadual
#> 110 2011                E   8  5080.939    Parda              Estadual
#> 111 2011                F   1 13711.338 Indígena               Federal
#> 112 2011                F   2 13330.930   Branca               Federal
#> 113 2011                F   4  9342.937    Preta               Federal
#> 114 2011                F   6 14711.613  Amarela               Federal
#> 115 2011                F   8 10247.837    Parda               Federal
#> 116 2011                M   1  2568.519 Indígena             Municipal
#> 117 2011                M   2  3283.735   Branca             Municipal
#> 118 2011                M   4  2532.473    Preta             Municipal
#> 119 2011                M   6  2868.355  Amarela             Municipal
#> 120 2011                M   8  2535.799    Parda             Municipal
#> 121 2012                E   1  7621.865 Indígena              Estadual
#> 122 2012                E   2  5894.413   Branca              Estadual
#> 123 2012                E   4  4956.461    Preta              Estadual
#> 124 2012                E   6  5966.544  Amarela              Estadual
#> 125 2012                E   8  5492.872    Parda              Estadual
#> 126 2012                F   1 13025.305 Indígena               Federal
#> 127 2012                F   2 13020.066   Branca               Federal
#> 128 2012                F   4  9222.460    Preta               Federal
#> 129 2012                F   6 14433.591  Amarela               Federal
#> 130 2012                F   8 10182.768    Parda               Federal
#> 131 2012                M   1  2849.622 Indígena             Municipal
#> 132 2012                M   2  3508.169   Branca             Municipal
#> 133 2012                M   4  2729.405    Preta             Municipal
#> 134 2012                M   6  3101.048  Amarela             Municipal
#> 135 2012                M   8  2758.311    Parda             Municipal
#> 136 2013                E   1  7820.534 Indígena              Estadual
#> 137 2013                E   2  6125.394   Branca              Estadual
#> 138 2013                E   4  5212.801    Preta              Estadual
#> 139 2013                E   6  6228.424  Amarela              Estadual
#> 140 2013                E   8  5746.524    Parda              Estadual
#> 141 2013                F   1 12493.242 Indígena               Federal
#> 142 2013                F   2 12445.097   Branca               Federal
#> 143 2013                F   4  8510.981    Preta               Federal
#> 144 2013                F   6 14379.355  Amarela               Federal
#> 145 2013                F   8  9695.658    Parda               Federal
#> 146 2013                M   1  3008.026 Indígena             Municipal
#> 147 2013                M   2  3639.552   Branca             Municipal
#> 148 2013                M   4  2832.762    Preta             Municipal
#> 149 2013                M   6  3135.173  Amarela             Municipal
#> 150 2013                M   8  2867.811    Parda             Municipal
#> 151 2014                E   1  8121.391 Indígena              Estadual
#> 152 2014                E   2  6533.967   Branca              Estadual
#> 153 2014                E   4  5547.861    Preta              Estadual
#> 154 2014                E   6  6685.271  Amarela              Estadual
#> 155 2014                E   8  6106.344    Parda              Estadual
#> 156 2014                F   1 12281.964 Indígena               Federal
#> 157 2014                F   2 12515.181   Branca               Federal
#> 158 2014                F   4  8490.974    Preta               Federal
#> 159 2014                F   6 14391.057  Amarela               Federal
#> 160 2014                F   8  9822.210    Parda               Federal
#> 161 2014                M   1  3109.209 Indígena             Municipal
#> 162 2014                M   2  3751.880   Branca             Municipal
#> 163 2014                M   4  2956.187    Preta             Municipal
#> 164 2014                M   6  3242.579  Amarela             Municipal
#> 165 2014                M   8  2958.452    Parda             Municipal
#> 166 2015                E   1  8163.123 Indígena              Estadual
#> 167 2015                E   2  6675.871   Branca              Estadual
#> 168 2015                E   4  5703.478    Preta              Estadual
#> 169 2015                E   6  6822.410  Amarela              Estadual
#> 170 2015                E   8  6251.601    Parda              Estadual
#> 171 2015                F   1 12289.452 Indígena               Federal
#> 172 2015                F   2 12595.075   Branca               Federal
#> 173 2015                F   4  8832.308    Preta               Federal
#> 174 2015                F   6 14368.898  Amarela               Federal
#> 175 2015                F   8 10096.656    Parda               Federal
#> 176 2015                M   1  3166.007 Indígena             Municipal
#> 177 2015                M   2  3727.451   Branca             Municipal
#> 178 2015                M   4  2957.306    Preta             Municipal
#> 179 2015                M   6  3253.814  Amarela             Municipal
#> 180 2015                M   8  2953.259    Parda             Municipal
#> 181 2016                E   1  7647.167 Indígena              Estadual
#> 182 2016                E   2  6465.657   Branca              Estadual
#> 183 2016                E   4  5460.763    Preta              Estadual
#> 184 2016                E   6  6534.610  Amarela              Estadual
#> 185 2016                E   8  5974.374    Parda              Estadual
#> 186 2016                F   1 11948.697 Indígena               Federal
#> 187 2016                F   2 11508.821   Branca               Federal
#> 188 2016                F   4  7866.175    Preta               Federal
#> 189 2016                F   6 13328.684  Amarela               Federal
#> 190 2016                F   8  9221.476    Parda               Federal
#> 191 2016                M   1  3123.696 Indígena             Municipal
#> 192 2016                M   2  3755.510   Branca             Municipal
#> 193 2016                M   4  2975.671    Preta             Municipal
#> 194 2016                M   6  3275.813  Amarela             Municipal
#> 195 2016                M   8  2948.033    Parda             Municipal
#> 196 2017                E   1  7568.603 Indígena              Estadual
#> 197 2017                E   2  6587.441   Branca              Estadual
#> 198 2017                E   4  5503.099    Preta              Estadual
#> 199 2017                E   6  6421.980  Amarela              Estadual
#> 200 2017                E   8  5965.173    Parda              Estadual
#> 201 2017                F   1 12635.362 Indígena               Federal
#> 202 2017                F   2 12550.452   Branca               Federal
#> 203 2017                F   4  8917.340    Preta               Federal
#> 204 2017                F   6 14446.971  Amarela               Federal
#> 205 2017                F   8 10089.474    Parda               Federal
#> 206 2017                M   1  3184.278 Indígena             Municipal
#> 207 2017                M   2  3832.140   Branca             Municipal
#> 208 2017                M   4  3032.050    Preta             Municipal
#> 209 2017                M   6  3353.941  Amarela             Municipal
#> 210 2017                M   8  2993.561    Parda             Municipal
#> 211 2018                E   1  7832.811 Indígena              Estadual
#> 212 2018                E   2  6853.642   Branca              Estadual
#> 213 2018                E   4  5720.761    Preta              Estadual
#> 214 2018                E   6  6849.534  Amarela              Estadual
#> 215 2018                E   8  6276.105    Parda              Estadual
#> 216 2018                F   1 11923.546 Indígena               Federal
#> 217 2018                F   2 12780.758   Branca               Federal
#> 218 2018                F   4  9136.230    Preta               Federal
#> 219 2018                F   6 14033.204  Amarela               Federal
#> 220 2018                F   8 10261.076    Parda               Federal
#> 221 2018                M   1  3221.001 Indígena             Municipal
#> 222 2018                M   2  3882.029   Branca             Municipal
#> 223 2018                M   4  3084.014    Preta             Municipal
#> 224 2018                M   6  3404.421  Amarela             Municipal
#> 225 2018                M   8  3038.765    Parda             Municipal
#> 226 2019                E   1  7865.962 Indígena              Estadual
#> 227 2019                E   2  6930.731   Branca              Estadual
#> 228 2019                E   4  5797.568    Preta              Estadual
#> 229 2019                E   6  6933.164  Amarela              Estadual
#> 230 2019                E   8  6420.943    Parda              Estadual
#> 231 2019                F   1 14867.021 Indígena               Federal
#> 232 2019                F   2 14788.049   Branca               Federal
#> 233 2019                F   4 10444.730    Preta               Federal
#> 234 2019                F   6 15479.554  Amarela               Federal
#> 235 2019                F   8 11427.856    Parda               Federal
#> 236 2019                M   1  3257.281 Indígena             Municipal
#> 237 2019                M   2  3926.008   Branca             Municipal
#> 238 2019                M   4  3120.837    Preta             Municipal
#> 239 2019                M   6  3456.694  Amarela             Municipal
#> 240 2019                M   8  3042.222    Parda             Municipal
#> 241 2020                E   1  7415.117 Indígena              Estadual
#> 242 2020                E   2  6685.455   Branca              Estadual
#> 243 2020                E   4  5633.151    Preta              Estadual
#> 244 2020                E   6  6715.285  Amarela              Estadual
#> 245 2020                E   8  6088.114    Parda              Estadual
#> 246 2020                F   1 11403.846 Indígena               Federal
#> 247 2020                F   2 12576.543   Branca               Federal
#> 248 2020                F   4  8662.912    Preta               Federal
#> 249 2020                F   6 14320.295  Amarela               Federal
#> 250 2020                F   8 10219.289    Parda               Federal
#> 251 2020                M   1  3315.191 Indígena             Municipal
#> 252 2020                M   2  3915.305   Branca             Municipal
#> 253 2020                M   4  3133.452    Preta             Municipal
#> 254 2020                M   6  3468.443  Amarela             Municipal
#> 255 2020                M   8  3072.137    Parda             Municipal
#> 256 2021                E   1  6869.472 Indígena              Estadual
#> 257 2021                E   2  6189.442   Branca              Estadual
#> 258 2021                E   4  5217.421    Preta              Estadual
#> 259 2021                E   6  6328.858  Amarela              Estadual
#> 260 2021                E   8  5661.047    Parda              Estadual
#> 261 2021                F   1 11055.484 Indígena               Federal
#> 262 2021                F   2 13420.173   Branca               Federal
#> 263 2021                F   4  9562.790    Preta               Federal
#> 264 2021                F   6 14345.523  Amarela               Federal
#> 265 2021                F   8 10441.602    Parda               Federal
#> 266 2021                M   1  3186.184 Indígena             Municipal
#> 267 2021                M   2  3681.942   Branca             Municipal
#> 268 2021                M   4  2961.480    Preta             Municipal
#> 269 2021                M   6  3235.979  Amarela             Municipal
#> 270 2021                M   8  2909.474    Parda             Municipal
if (FALSE) { # \dontrun{
get_series(series_id = 230:232)
} # }