Search for specific words

The search_series() function provides a quick way to access a list of series that may be of interest, along with their respective IDs. Each series appears according to the number of themes it is present in.

If one result should appear before the other, the result can be sorted by using the word multiple times in the search, thus giving it greater importance.

Usage

search_series(words, case_insensitive = TRUE, require_all = FALSE)

Arguments

words

A word or an array of words to check

case_insensitive

Ignore the difference between uppercase and lowercase

require_all

Require all words

Value

A data.frame

Examples

search_series("remuneração")
#>                                                                                                         series_title
#> 1              Remuneração média mensal no setor público, nos Poderes e nos níveis federativos, por sexo (1986-2021)
#> 2                                                          Remuneração média e mediana mensal, por setor (1986-2017)
#> 3                                      Remuneração média e mediana mensal, por Poder ou nível federativo (1986-2021)
#> 4                                   Remuneração média e mediana mensal nos Poderes, por nível federativo (1986-2021)
#> 5            Remuneração média mensal no Brasil, nas regiões e nos estados, por Poder e nível federativo (1985-2021)
#> 6                    Remuneração média mensal no Brasil, nas regiões e nos estados, por nível federativo (1985-2019)
#> 7                                                     Distribuição de remuneração nos níveis federativos (1985-2019)
#> 8                                                                Distribuição de remuneração nos Poderes (1985-2019)
#> 9                                   Remuneração média mensal nos níveis federativos ou Poderes, por sexo (1986-2017)
#> 10                                  Remuneração média mensal nos níveis federativos ou Poderes, por sexo (1986-2017)
#> 11                                  Remuneração média mensal nos níveis federativos ou Poderes, por sexo (1986-2017)
#> 12                           Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 13                           Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 14                           Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 15            Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 16            Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 17            Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 18            Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 19                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 20                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 21                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 22                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 23                             Distribuição de remuneração nos Poderes segredados por níveis federativos (1985-2019)
#> 24                              Remuneração média mensal no Brasil, nas regiões e nos estados, por Poder (1985-2019)
#> 25 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#> 26 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#> 27 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#> 28                            Remuneração média no Brasil, regiões e estados por setor público e privado (1985-2019)
#> 29                                               Remuneração média mensal de vínculos públicos por cor (2004 - 2021)
#> 30                                               Remuneração média mensal de vínculos públicos por cor (2004 - 2021)
#> 31                                    Remuneração média mensal de vínculos públicos nos Poderes, por cor (2004-2021)
#> 32                                    Remuneração média mensal de vínculos públicos nos Poderes, por cor (2004-2021)
#> 33                                    Remuneração média mensal de vínculos públicos nos Poderes, por cor (2004-2021)
#> 34                         Remuneração média mensal de vínculos públicos nos níveis federativos, por cor (2004-2021)
#> 35                         Remuneração média mensal de vínculos públicos nos níveis federativos, por cor (2004-2021)
#> 36                         Remuneração média mensal de vínculos públicos nos níveis federativos, por cor (2004-2021)
#> 37                                          Remuneração média mensal de vínculos públicos por sexo e cor (2004-2021)
#> 38                                          Remuneração média mensal de vínculos públicos por sexo e cor (2004-2021)
#>    series_id                                        theme_title theme_id
#> 1         94                      Escolaridade no setor público       35
#> 2        114                      Remunerações no setor público       26
#> 3        115                      Remunerações no setor público       26
#> 4        116                      Remunerações no setor público       26
#> 5        117                      Remunerações no setor público       26
#> 6        118                      Remunerações no setor público       26
#> 7        119                      Remunerações no setor público       26
#> 8        120                      Remunerações no setor público       26
#> 9        140                      Remunerações no setor público       26
#> 10       140 Vínculos e remunerações no setor público, por sexo       28
#> 11       140                     Desigualdades no setor público       49
#> 12       143                      Remunerações no setor público       26
#> 13       143                 Funcionalismo público por cor/raça       41
#> 14       143                     Desigualdades no setor público       49
#> 15       144                      Remunerações no setor público       26
#> 16       144                      Escolaridade no setor público       35
#> 17       144                 Funcionalismo público por cor/raça       41
#> 18       144                     Desigualdades no setor público       49
#> 19       145                      Remunerações no setor público       26
#> 20       145 Vínculos e remunerações no setor público, por sexo       28
#> 21       145                 Funcionalismo público por cor/raça       41
#> 22       145                     Desigualdades no setor público       49
#> 23       173                      Remunerações no setor público       26
#> 24       174                      Remunerações no setor público       26
#> 25       197 Vínculos públicos por Poderes e níveis federativos       25
#> 26       197                      Remunerações no setor público       26
#> 27       197 Vínculos e remunerações no setor público, por sexo       28
#> 28       200                      Remunerações no setor público       26
#> 29       227                      Remunerações no setor público       26
#> 30       227                 Funcionalismo público por cor/raça       41
#> 31       229 Vínculos públicos por Poderes e níveis federativos       25
#> 32       229                      Remunerações no setor público       26
#> 33       229                 Funcionalismo público por cor/raça       41
#> 34       230 Vínculos públicos por Poderes e níveis federativos       25
#> 35       230                      Remunerações no setor público       26
#> 36       230                 Funcionalismo público por cor/raça       41
#> 37       240                      Remunerações no setor público       26
#> 38       240                 Funcionalismo público por cor/raça       41
search_series(c("remuneração", "raça"), require_all = TRUE)
#>                                                                                                         series_title
#> 1                            Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 2                            Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 3                            Remuneração líquida média mensal no Executivo civil federal ativo, por raça (1999-2020)
#> 4             Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 5             Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 6             Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 7             Remuneração líquida média mensal no Executivo civil federal ativo, por raça e escolaridade (1999-2020)
#> 8                     Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 9                     Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 10                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 11                    Remuneração líquida média mensal no Executivo civil federal ativo, por sexo e raça (1999-2020)
#> 12 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#> 13 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#> 14 Distribuição da remuneração média e dos vínculos de trabalho por raça, sexo, Poder e nível federativo (2004-2020)
#>    series_id                                        theme_title theme_id
#> 1        143                      Remunerações no setor público       26
#> 2        143                 Funcionalismo público por cor/raça       41
#> 3        143                     Desigualdades no setor público       49
#> 4        144                      Remunerações no setor público       26
#> 5        144                      Escolaridade no setor público       35
#> 6        144                 Funcionalismo público por cor/raça       41
#> 7        144                     Desigualdades no setor público       49
#> 8        145                      Remunerações no setor público       26
#> 9        145 Vínculos e remunerações no setor público, por sexo       28
#> 10       145                 Funcionalismo público por cor/raça       41
#> 11       145                     Desigualdades no setor público       49
#> 12       197 Vínculos públicos por Poderes e níveis federativos       25
#> 13       197                      Remunerações no setor público       26
#> 14       197 Vínculos e remunerações no setor público, por sexo       28